WIP: Refactoring flow parser
