+++ src/lib/converter/helper/actionParser.ts
@@ -1,8 +1,8 @@
-import { ActionCall, InputParamValue } from '../types/flow';
-import { RecordCreate, RecordUpdate, RecordFilter, RecordLookup } from '../types/flowRecordAction';
-import { toArray } from './util/arrayUtils';
-import { ProcessMetadataValue } from '../types/processMetadataValue';
-import { ReadableCondition, ReadableActionItem, ReadableActionItemParameter } from '../types/parser';
+import { ActionCall, InputParamValue } from '../../../types/metadata/flow';
+import { RecordCreate, RecordUpdate, RecordFilter, RecordLookup } from '../../../types/metadata/flowRecordAction';
+import { toArray } from '../../util/arrayUtils';
+import { ProcessMetadataValue } from '../../../types/metadata/processMetadataValue';
+import { ReadableCondition, ReadableActionItem, ReadableActionItemParameter } from '../../../types/converter';
 
 const layout = require('./actionLayout.json');
